Implement post API endpoints
The defroutes-api macro does a lot of the heavy lifting here. My reasoning is
that the api route definitions would constantly be repeating "/api/" and
repeatedly parsing json from the request body for post requests and generating
json responses. The macro takes care of all of that so that the route handler
functions can focus on unique logic.

With the scope of this project, it will probably turn out to be more code than
if I had repeated the parsing/generation, but for a more complex api it could
save a lot of time and cut down on loc. I also probably could've used some ring
middleware, but I wanted to get more practice writing macros.