Rename project to org-blog
I think it probably makes sense to keep the blog/portfolio server as a separate
project from the frontend. This will go along with the emacs mode that I plan to
write to allow people to post org-mode blog posts directly to the server.

For my own site, I'll set up a simple fullstack clojurescript app and then proxy
api requests to the blog server.